Move BotAction into core/poker; make the fallbacks engine-private
BotAction and two fallback constants lived on the AI provider port. BotAction was
there for a defensible reason -- it is AIService.query()'s return type, so the DTO
sat next to its port -- but it outgrew that, and the constants never belonged.

BotAction is a poker concept, not an AI-client one. The live action-executor
performs one at the table and the arena agent maps one onto the engine's legal
actions; neither cares that a model produced it. The symptom was live/ importing
from core/ai/ just to name a decision. core/poker already holds this kind of
shared vocabulary -- the Action enum was lifted there for the same reason -- so
BotAction moves to core/poker/bot-action.ts and live/arena now import a poker
type from the poker layer.

defaultCheckAction/defaultFoldAction encode poker policy ("check if legal, else
fold") and had exactly one consumer, DecisionEngine.fallback(), yet were exported
from the provider port. They are now module-private to decision-engine.

Two latent bugs closed while moving them:

- They were untyped, so inferred structurally. A typo in action_str compiled and
  only surfaced at runtime as an unexplained fold. Now annotated : BotAction.
- They were shared mutable singletons whose identity escapes into live and arena
  (fallback returns the object, not a copy). Nothing mutates them today, but one
  future write would corrupt every later fallback in the process. Now frozen, so
  such a write throws instead of silently poisoning subsequent decisions.

No re-export shim on the old path -- that would preserve the ambiguity this
removes -- so all 8 import sites were updated. Pure move plus typing/freezing; no
behavior change intended.
