+++ src/core/ai/ai-query.helper.ts
@@ -1,4 +1,4 @@
-import { BotAction } from "./ai-client.interface.ts";
+import { BotAction } from "../poker/bot-action.ts";
 
 export const playstyleToPrompt: Map<string, string> = new Map<string, string>([
     ["pro", "You are a pro poker player who plays strong ranges preflop and plays aggressively postflop."],
