+++ src/core/ai/openai.service.ts
@@ -1,6 +1,7 @@
 import OpenAI from "openai";
 import { ChatCompletionCreateParamsNonStreaming, ChatCompletionMessageParam } from "openai/resources/chat/completions.mjs";
-import { AIService, BotAction } from "./ai-client.interface.ts";
+import { AIService } from "./ai-client.interface.ts";
+import { BotAction } from "../poker/bot-action.ts";
 import { getPromptFromPlaystyle, parseResponse} from "./ai-query.helper.ts";
 import { withTimeout } from "../../utils/bot-timeout.helper.ts";
